requirements.txt

streamlit
transformers
torch
faiss-cpu
sentence-transformers
pandas
